trying to fix bug
